tests/wpt/web-platform-tests/css/css-ui/parsing/outline-valid-mandatory.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing outline with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline">
<link rel="help" href="https://drafts.csswg.org/cssom/#serializing-css-values">
<meta name="assert" content="outline supports the full grammar '<outline-color> || <outline> || <outline>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("outline", "rgba(10, 20, 30, 0.4)");

test_valid_value("outline", "auto");
test_valid_value("outline", "none", ["none", "medium"]); // Edge serializes as "invert"
test_valid_value("outline", "dotted");
test_valid_value("outline", "dashed");
test_valid_value("outline", "solid");
test_valid_value("outline", "double");
test_valid_value("outline", "groove");
test_valid_value("outline", "ridge");
test_valid_value("outline", "inset");
test_valid_value("outline", "outset");

test_valid_value("outline", "0", "0px");
test_valid_value("outline", "1px");
test_valid_value("outline", "calc(2em + 3ex)");
test_valid_value("outline", "thin");
test_valid_value("outline", "medium", ["none", "medium"]); // Edge serializes as "invert"
test_valid_value("outline", "thick");

test_valid_value("outline", "dashed thin");
test_valid_value("outline", "medium rgba(10, 20, 30, 0.4)", ["rgba(10, 20, 30, 0.4) medium", "rgba(10, 20, 30, 0.4)"]);

test_valid_value("outline", "3px ridge rgba(10, 20, 30, 0.4)", "rgba(10, 20, 30, 0.4) ridge 3px");
</script>
</body>
</html>
